presentations/test-the-web-forward/index.html

<!doctype html>  
<html lang="en">
	
	<head>
		<meta charset="utf-8">
		
		<title>W3C Test Writing Tutorials</title>

		<meta name="description" content="An easy to use CSS 3D slideshow tool for quickly creating good looking HTML presentations.">
		<meta name="author" content="Hakim El Hattab">
		
		<link href='http://fonts.googleapis.com/css?family=Lato:400,700,400italic,700italic' rel='stylesheet' type='text/css'>
		
		<link rel="stylesheet" href="css/reset.css">
		<link rel="stylesheet" href="css/main.css">
		<link rel="stylesheet" href="css/adobe.css">
		<link rel="stylesheet" href="lib/zenburn.css">
	</head>
	
	<body>
		
		<div id="reveal">
			<!-- Any section element inside of this container is displayed as a slide -->
			<div class="slides">
			    <script type="text/javascript">
  			    function computedStyle(element)
          	{
          	  document.getElementById("computedStyle").innerHTML = window.getComputedStyle(element).getPropertyValue("-webkit-transform");
          	}

            function runAnimation(id)
          	{
          	   var shape = document.getElementById(id);
          	   var state = shape.style.webkitAnimationPlayState;
          	   
          	   if(state == "running")
          	        shape.style.webkitAnimationPlayState = "paused";
          	  else
          	        shape.style.webkitAnimationPlayState = "running";
          	   
          	}
          </script>
                <section>
                   	<h2>Test The Web Forward</h2>
                   	<img src="assets/images/W3C_logo.jpg">
                   </section>

                   <section id="Agenda">
                       <h2>Agenda</h2>
                       <ul>
                           <li>Mercurial</li>
                           <li>Ref Test Example</li>
                           <li>JavaScript Test Example</li>
                           <li>Q&A</li>
                   </section>
				<section>
					<h2>W3C Reftest Tutorial</h2>
					<h3>Sample Test Case</h3>
					<br>
					<h4>Write a basic test for the SVG transform attribute described in the CSS3 Transforms spec using the translate() function</h4>
					<br>
					<h3>Spec Links:</h3>
					<br>
				    <h4>
					<a href="http://dev.w3.org/csswg/css3-transforms/#svg-transform" target="top">http://dev.w3.org/csswg/css3-transforms/#svg-transform</a>
					<br>
					<br>
					<a href="http://dev.w3.org/csswg/css3-transforms/#two-d-transform-functions" target="top">http://dev.w3.org/csswg/css3-transforms/#two-d-transform-functions</a>
					</h4>
					
					<script>
						// Delicously hacky. Look away.
						if( navigator.userAgent.match( /(iPhone|iPad|iPod|Android)/i ) )
						document.write( '<p style="color: rgba(0,0,0,0.3); text-shadow: none;">('+'Tap to navigate'+')</p>' );
					</script>
				</section>
			    <section>
			        <h2>Step 1</h2>
			        <br>
    			    <h3>Pull the starter files for this tutorial from Mercurial</h3>
			        
			        <p>OSX:<p>
  				    <div class="console">
  				    <div class="consoleText">
  				        <br><br>
  			            <p>$ cd [...]/W3Cdir/test</p>
  			            <p>$ hg pull -u</p>
  			            <p>$ ls -lR contributors/adobe/ttwf</p>
  			        </div>
  			        </div>
  			        <div class="bottomNote">
  			            <h5>
  			                Note: These files can also be downloaded from here:
  			                <br>
  			                <a href="assets/resources/ttwf-reftest-tutorial.zip">ttwf-reftest-tutorial.zip</a>
  			            </h5>
  			        </div>
  			        <br><br><br><br><br><br><br>
  			        <p>Windows:<p>
  			         <div class="console">
      				    <div class="consoleText">
      				        <br><br>
      			            <p>> cd [...]\W3Cdir\test</p>
      			            <p>> hg pull -u</p>
      			            <p>> dir contributors\adobe\ttwf</p>
      			        </div>
      			    </div>
                </section>
                <section>
    			     <h2>Step 2</h2>
    			     <br>
        		     <h3>Create a directory for your new test</h3>
    			     <br>
    			     <p>OSX:<p>
      				 <div class="console">
      				 <div class="consoleText">
      				     <br>
      				     <br>
      				     <br>
      				     <p>$ cd [...]/W3Cdir/test/contributors/adobe/ttwf/</p>
      				     <p>$ mkdir -p yourName/incoming</p>
      			     </div>
      			     </div>
      			     <br><br><br><br><br>
      			     <p>Windows:<p>
       				 <div class="console">
       				 <div class="consoleText">
       				     <br>
       				     <br>
       				     <br>
       				     <p>> cd [...]\W3Cdir\test\contributors\adobe\ttwf</p>
       				     <p>> mkdir yourName\incoming</p>
       			     </div>
       			     </div>
                </section>
                <section>
    			     <h2>Step 3</h2>
    			     <br>
        		     <h3>Copy the starter files to your incoming directory</h3>
    			     <br>
     			     <p>OSX:<p>
      				 <div class="console">
      				 <div class="consoleText">
      				     <br>
      				     <br>
      				     <p>$ cd [...]/W3Cdir/test/contributors/adobe/ttwf/yourName/incoming</p>
      			         <p>$ cp -R [...]/W3Cdir/test/contributors/adobe/ttwf/* .</p>
      			     </div>
      			     </div>
      			     <br><br><br><br><br>
      			     <p>Windows:<p>
       				 <div class="console">
       				 <div class="consoleText">
       				     <p>> cd [...]\W3Cdir\test\contributors\adobe\ttwf\yourName\incoming</p>
       			         <p>> copy ..\..\* .</p>
       			         <p>> mkdir reftest</p>
       			         <p>> cd reftest</p>
       			         <p>> copy ..\..\..\reftest\* .</p>
       			     </div>
       			     </div>
                </section>
                <section>
    			     <h2>Step 4</h2>
    			     <h3>Open the test file in an editor</h3>
    			     <img src="assets/images/testStarter.png">
                </section>
                <section>
    			     <h2>Step 5</h2>
    			     <br>
        		     <h3>Add the test metadata</h3>
    			     <div class="tab5">
    			        <ul>
      				        <li>Title</li>
      				        <li>Author</li>
      				        <li>Spec link(s)</li>
      				        <li>Reference file path</li>
      				        <li>Flags</li>
      				        <li>Assertion</li>
      				    </ul>
      				</div>
                </section>
                <section>
    			     <h2>Step 6</h2>
    			     <br>
        		     <h3>Add the transform attribute and test function to the SVG element</h3>
        		     <br>
        		     <div class="editor">
       				 <div class="editorText">
       				     <br><br>
       				     <p class="tab1 editorHtml">&lt;body&gt;
       				     <p class="tab2 editorHtml">&lt;svg&gt;</p>
       			         <p class="tab3 editorHtml">&lt;rect width=<span class="editorAttrVal">"100"</span>
       			             <span class="editorHtml">height=</span><span class="editorAttrVal">"100"</span>
       			             <span class="highlight"><span class="editorHtml">transform=</span><span class="editorAttrVal">"translate(50)"</span></span><span class="editorHtml">/&gt;</span>
       			         </p>
       			         <p class="tab2 editorHtml">&lt;/svg&gt;</p>
       			         <p class="tab1 editorHtml">&lt;/body&gt;
       			     </div>
       			     </div>
    		    </section>
                <section>
    			     <h2>Step 7</h2>
    			     <br>
        		     <h3>Create an element that is visible only upon failure</h3>
    			     <br>
         		     <div class="editor">
        			 <div class="editorText">
            		     <br>
            		     <p class="tab1 editorHtml">&lt;body&gt;
            		     <p class="tab2 editorHtml">&lt;svg&gt;</p>
            		     <p class="tab3 editorHtml"><span class="highlight">&lt;rect x=<span class="editorAttrVal">"51"</span>
              		         <span class="editorHtml">y=</span><span class="editorAttrVal">"1"<span>
              		         <span class="editorHtml">width=</span><span class="editorAttrVal">"98"</span>
              		         <span class="editorHtml">height=</span><span class="editorAttrVal">"98"</span><span class="editorHtml">/&gt;</span></span>
              		     </p>
              		     <p class="tab3 editorHtml">&lt;rect width=<span class="editorAttrVal">"100"</span>
            		         <span class="editorHtml">height=</span><span class="editorAttrVal">"100"</span>
            		         <span class="editorHtml">transform=</span><span class="editorAttrVal">"translate(50)"<span><span class="editorHtml">/&gt;</span>
            		     </p>
            		     <p class="tab2 editorHtml">&lt;/svg&gt;</p>
            			 <p class="tab1 editorHtml">&lt;/body&gt;
            		 </div>
        			 </div>
                </section>
                <section>
    			     <h2>Step 8</h2>
    			     <br>
        		     <h3>Add green & red fill colors to denote passing & failing conditions</h3>
        		     <br>
        		     <div class="editor">
         			 <div class="editorText">
         			     <p class="tab1 editorHtml">&lt;body&gt;
             		     <p class="tab2 editorHtml">&lt;svg&gt;</p>
             		     <p class="tab3 editorHtml">&lt;rect width=<span class="editorHtml">x=</span><span class="editorAttrVal">"51"<span>
               		         <span class="editorHtml">y=</span><span class="editorAttrVal">"1"<span>
               		         <span class="editorHtml">width=</span><span class="editorAttrVal">"98"</span>
               		         <span class="editorHtml">height=</span><span class="editorAttrVal">"98"</span>
               		         <span class="highlight"><span class="editorHtml">fill=</span><span class="editorAttrVal">"red"</span></span><span class="editorHtml">/&gt;</span>
               		     </p>
             		     <p class="tab3 editorHtml">&lt;rect width=<span class="editorAttrVal">"100"</span>
             		         <span class="editorHtml">height=</span><span class="editorAttrVal">"100"</span>
             		         <span class="editorHtml">transform=</span><span class="editorAttrVal">"translate(50)"<span>
             		         <span class="editorHtml"><span class="highlight">fill=<span class="editorAttrVal">"green"</span></span><span class="editorHtml">/&gt;</span>
             		     </p>
             		     <p class="tab2 editorHtml">&lt;/svg&gt;</p>
             			 <p class="tab1 editorHtml">&lt;/body&gt;
             		 </div>
         			 </div>
    			</section>
                <section>
    			     <h2>Step 9</h2>
    			     <br>
        		     <h3>Add a statement to the test page to make this a self-describing test</h3>
        		     <br>
        		     <div class="editor">
         			 <div class="editorText">
         			     <p class="tab1 editorHtml">&lt;body&gt;
             		     <p class="tab2 editorHtml"><span class="highlight">&lt;p&gt;<span class="editorHtmlTxt">The test passes if there is a green square and no red.</span><span class="editorHtml">&lt;/p&gt;</span></span>
                         <p class="tab2 editorHtml">&lt;svg&gt;</p>
             		     <p class="tab3 editorHtml">&lt;rect x=</span><span class="editorAttrVal">"51"<span>
               		         <span class="editorHtml">y=</span><span class="editorAttrVal">"1"<span>
               		         <span class="editorHtml">width=</span><span class="editorAttrVal">"98"</span>
               		         <span class="editorHtml">height=</span><span class="editorAttrVal">"98"</span>
               		         <span class="editorHtml">fill=</span><span class="editorAttrVal">"red"</span><span class="editorHtml">/&gt;</span>
               		     </p>
             		     <p class="tab3 editorHtml">&lt;rect width=<span class="editorAttrVal">"100"</span>
             		         <span class="editorHtml">height=</span><span class="editorAttrVal">"100"</span>
             		         <span class="editorHtml">transform=</span><span class="editorAttrVal">"translate(50)"<span>
             		         <span class="editorHtml">fill=</span><span class="editorAttrVal">"green"</span><span class="editorHtml">/&gt;</span>
             		     </p>
             		     <p class="tab2 editorHtml">&lt;/svg&gt;</p>
             			 <p class="tab1 editorHtml">&lt;/body&gt;
             		 </div>
         			 </div>
    			</section>
                <section>
    			     <h2>Step 10</h2>
    			     <br>
        		     <h3>Open the reference file in an editor</h3>
        		     <img src="assets/images/refStarter.png">
    		    </section>
    		    <section>
    			     <h2>Step 10</h2>
    			     <br>
        		     <h3>Add the reference metadata</h3>
    			     <div class="tab5">
    			        <ul>
      				        <li>Title</li>
      				        <li>Author</li>
      				        <li>Flags</li>
      				    </ul>
      				</div>
                </section>
                <section>
        		     <h2>Step 11</h2>
        		     <br>
            	     <h3>Create and position the reference element so it matches the test file exactly in a passing condition</h3>
            	     <br>
            	        <div class="editor">
        				<div class="editorText">
        				     <br><br>
        				     <p class="tab1 editorHtml">&lt;body&gt;
        				     <p class="tab2 editorHtml">&lt;svg&gt;</p>
        			         <p class="tab3 editorHtml"><span class="highlight">&lt;rect x=<span class="editorAttrVal">"50"</span>
        			             <span class="editorHtml">y=</span><span class="editorAttrVal">"0"<span>
                                 <span class="editorHtml">width=</span><span class="editorAttrVal">"100"<span>
                                 <span class="editorHtml">height=</span><span class="editorAttrVal">"100"</span>
        			             <span class="editorHtml">fill=</span><span class="editorAttrVal">"green"</span><span class="editorHtml">/&gt;</span></span>
        			         </p>
        			         <p class="tab2 editorHtml">&lt;/svg&gt;</p>
        			         <p class="tab1 editorHtml">&lt;/body&gt;
        			    </div>
        			    </div>
                </section>
                <section>
        		     <h2>Step 12</h2>
        		     <br>
            	     <h3>Add the self-describing statement here, too</h3>
            	     <br>
            	     <div class="editor">
     				 <div class="editorText">
     				     <br><br>
     				     <p class="tab1 editorHtml">&lt;body&gt;
     				     <p class="tab2 editorHtml"><span class="highlight">&lt;p&gt;<span class="editorHtmlTxt">The test passes if there is a green square and no red.</span><span class="editorHtml">&lt;/p&gt;</span></span>
                         <p class="tab2 editorHtml">&lt;svg&gt;</p>
     			         <p class="tab3 editorHtml">&lt;rect x=<span class="editorAttrVal">"50"</span>
     			             <span class="editorHtml">y=</span><span class="editorAttrVal">"0"<span>
                              <span class="editorHtml">width=</span><span class="editorAttrVal">"100"<span>
                              <span class="editorHtml">height=</span><span class="editorAttrVal">"100"</span>
     			             <span class="editorHtml">fill=</span><span class="editorAttrVal">"green"</span><span class="editorHtml">/&gt;</span>
     			         </p>
     			         <p class="tab2 editorHtml">&lt;/svg&gt;</p>
     			         <p class="tab1 editorHtml">&lt;/body&gt;
     			    </div>
     			    </div>
                </section>
                <section>
        		     <h2>Step 13</h2>
        		     <br>
            	     <h3>Examine your test and reference files in the browser and confirm they are identical</h3>
            	     <img src="assets/images/passingRefTestExample.png">
                </section>
                <section>
        		     <h3>Failing test example</h3>
        		     <br>
            	     <img src="assets/images/failingRefTestExample.png">
                </section>
                <section>
    			     <h2>Step 14</h2>
    			     <br>
        		     <h3>Add your new test & reference file to the repository</h3>
    			     <br>
    			     <p>OSX:</p>
      				 <div class="console">
      				 <div class="consoleText">
      				     <br>
      				     <p>$ cd [...]/W3Cdir/test/contributors/adobe/ttwf/yourName/incoming</p>
      			         <p>$ hg add</p>
      			         <p>$ hg commit -m "TTWF reftest tutorial"</p>
      			         <p>$ hg push</p>
      			     </div>
      			     </div>
      			     <br><br><br><br><br>
      			      <p>Windows:</p>
       				 <div class="console">
       				 <div class="consoleText">
       				     <br>
       				     <p>> cd [...]\W3Cdir\test\contributors\adobe\ttwf\yourName\incoming</p>
       			         <p>> hg add</p>
       			         <p>> hg commit -m "TTWF reftest tutorial"</p>
       			         <p>> hg push</p>
       			     </div>
       			     </div>
                </section>
                <section>
    			     <h2>Step 15</h2>
    			     <br>
        		     <h3>Ask someone to review your test</h3>
    		    </section>
    		     <section>
        			     <h2>Step 16</h2>
        			     <br>
            		     <h3>Incorporate feedback & move test to the submitted folder</h3>
        			     <br>
          				 <div class="consoleLarge">
          				 <div class="consoleText">
          				     <p>$ cd ~/W3Cdir/test/contributors/yourName</p>
          				     <p>$ hg pull -u</p>
          				     <p>$ hg merge</p>
          				     <p>$ hg commit -m "merging"</p>
          			         <p>$ hg mv incoming/ttwf-reftest-tutorial-001.html submitted/ttwf-reftest-tutorial-001.html</p>
          			         <p>$ hg mv incoming/reftest/ttwf-reftest-tutorial-ref.html submitted/reftest/ttwf-reftest-tutorial-ref.html</p>
          			         <p>$ hg commit -m "moved the TTWF reftest tutorial to the submitted folder"</p>
          			         <p>$ hg push</p>
          			     </div>
          			     </div>
                    </section>
                
				<section>
					<h2>JavaScript Test Overview</h2>
					
					<ul>
					    <li>In most cases, JavaScript is the preferred method for writing tests</li>
					    <li>JavaScript tests have some advantages over Ref Tests
					        <ul>
					            <li>Robust</li>
					            <li>Flexible</li>
					            <li>Performant</li>
					            <li>Powerful</li>
					        </ul>
					    </li>
					    <li>W3C provides a framework (testharness) to simplify and standardize test creation
					        <ul>
					            <li>Includes an API for making common test assertions</li>
					            <li>Formats results and displays debug information for failures</li>
					        </ul>
					    </li>
					    <li>JavaScript tests are best when used in conjunction with some number of Ref Tests that verify final rendering</li>
					</ul>
				</section>
					
				<section>
				    <h2>W3C JavaScript Tutorial</h2>
					<h3>Sample Test Case</h3>
					<br>
					<h4>Write a basic test for the SVG transform attribute described in the CSS3 Transforms spec using the translate() function</h4>
					<h4 style="color:red">Same as Ref Test example</h4>
					<br>
					<h3>Spec Links:</h3>
					<br>
				    <h4>
					<a href="http://dev.w3.org/csswg/css3-transforms/#svg-transform" target="top">http://dev.w3.org/csswg/css3-transforms/#svg-transform</a>
					<br>
					<br>
					<a href="http://dev.w3.org/csswg/css3-transforms/#two-d-transform-functions" target="top">http://dev.w3.org/csswg/css3-transforms/#two-d-transform-functions</a>
					</h4>
					
					<script>
						// Delicously hacky. Look away.
						if( navigator.userAgent.match( /(iPhone|iPad|iPod|Android)/i ) )
						document.write( '<p style="color: rgba(0,0,0,0.3); text-shadow: none;">('+'Tap to navigate'+')</p>' );
					</script>
				</section>
				
			    <section>
			        <h2>Step 1</h2>
			        <br>
    			    <h3>Did you clone the resources directory?</h3>
    			    <div class="console">
    			        <div class="consoleText">
        			        <br><br>
        			        <p>$ ls /resources</p>
        			        <p>... testharness.css   testharness.js    testharnessreport.js</p>
        			    </div>
        			</div>
        			<br><br><br><br><br>
        			<h3>If not, clone the resources directory now:</h3>
                    <div class="console">
    			        <div class="consoleText">
        			        <br><br>
        			        <p>$ cd /</p>
        			        <p>$ sudo hg clone https://dvcs.w3.org/hg/resources/</p>
        			    </div>
        			</div>
                </section>
                				
			    <section>
			        <h2>Step 2</h2>
			        <br>
    			    <h3>JavaScript starter file was copied during Mercurial setup</h3>
    			    <h3>Open the test file in an editor</h3>
    			    <img src="assets/images/jsTestStarter.png">
                </section>
                
                <section>
    			     <h2>Step 3</h2>
    			     <br>
        		     <h3>Add the test metadata</h3>
    			     <div class="tab5">
    			        <ul>
      				        <li>Title</li>
      				        <li>Author</li>
      				    </ul>
      				</div>
                </section>
                
                <section>
        			<h2>Step 4</h2>
    			    <br>
        		    <h3>Import the W3C JavaScript Test Harness files</h3>                
                    <div class="editor">
                        <div class="editorText">
                            <br>
                            <p class="tab1 editorGray">&lt;meta name="assert" content="TEST ASSERTION"&gt;</p>
                            <p class="tab1 editorHtml">&lt;script src="/resources/testharness.js"&gt;&lt;/script&gt;</p>
                            <p class="tab1 editorHtml">&lt;script src="/resources/testharnessreport.js"&gt;&gt;&lt;/script&gt;</p>
                            <p class="tab1 editorGray">&lt;style&gt;</p>
                            <p class="tab2 editorGray">...</p>
                            <p class="tab1 editorGray">&lt;/style&gt;</p>
                        </div>
                    </div>
                </section>

                <section>
        			<h2>Step 5</h2>
    			    <br>
        		    <h3>Add element with id="log" for test results</h3>
        		    <br>
                    <div class="editor">
       				    <div class="editorText">
           				    <br><br>
           				    <p class="tab1 editorGray">&lt;body&gt;</p>
           				    <p class="tab2 editorGray">...</p>
           			        <p class="tab2 editorHtml">&lt;div id=<span class="editorAttrVal">"log"</span>&gt;&lt;/div&gt;</span>
           			        <p class="tab1 editorGray">&lt;/body&gt;
       			        </div>
       			     <br><br>
       			     <h3>testharness will add a HTML table that contains test results to this element</h3>
                </section>
                
                <section>
    			     <h2>Step 6</h2>
    			     <br>
        		     <h3>Add the test function to the SVG element</h3>
        		     <h4 style="color:red">Same as step 6 from RefTest section </h4>
        		     <br>
        		     <div class="editor">
       				     <div class="editorText">
           				     <br><br>
           				     <p class="tab1 editorHtml">&lt;body&gt;
           				     <p class="tab2 editorHtml">&lt;svg&gt;</p>
           			         <p class="tab3 editorHtml">&lt;rect width=<span class="editorAttrVal">"100"</span>
           			             <span class="editorHtml">height=</span><span class="editorAttrVal">"100"</span>
           			             <span class="highlight"><span class="editorHtml">transform=</span><span class="editorAttrVal">"translate(50,50)"</span></span><span class="editorHtml">/&gt;</span>
           			         </p>
           			         <p class="tab2 editorHtml">&lt;/svg&gt;</p>
           			         <p class="tab1 editorHtml">&lt;/body&gt;
           			     </div>
       			     </div>
    		    </section>
    		    
                <section>
    			     <h2>Step 7</h2>
    			     <br>
        		     <h3>Set margins and padding to 0px</h3>
        		     <br>
        		     <div class="editor">
       			         <div class="editorText">
           			         <br><br>
           				     <p class="tab1 editorHtml">&lt;style&gt;</p>
           				     <p class="tab2 editorHtml">* {</p>
           			         <p class="tab3 editorHtml">margin: 0px;</p>
           			         <p class="tab3 editorHtml">padding: 0px;</p>
           			         <p class="tab2 editorHtml">}</p>
           			         <p class="tab1 editorHtml">&lt;/style&gt;
           			     </div>
       			     </div>
       			     <br><br><br><br><br><br>
       			     <h3>This is not required, but simplifies the code</h3>
    		    </section>    		    
    		    
                <section>
    			     <h2>Step 8</h2>
    			     <br>
        		     <h3>Create a script block that will contain the JavaScript</h3>
    			     <br>
         		     <div class="editor">
        			     <div class="editorText">
                		     <p class="tab1 editorGray">&lt;body&gt;</p>
           				     <p class="tab2 editorGray">&lt;svg&gt;</p>
           			         <p class="tab3 editorGray">&lt;rect width="100"
           			             <span class="editorGray">height="100"</span>
           			             <span class="editorGray">transform="translate(50,50)"/&gt;</span>
           			         </p>
           			         <p class="tab2 editorGray">&lt;/svg&gt;</p>
               			     <p class="tab2 editorGray">&lt;div id="log"&gt;&lt;/div&gt;</p>        			            
                		     <p class="tab2 editorHtml">&lt;script&gt;</p>
                		     <p class="tab2 editorHtml">&lt;/script&gt;</p>
                			 <p class="tab1 editorGray">&lt;/body&gt;</p>
                		 </div>
        			 </div>
                </section>
                
                <section>
    			     <h2>Step 9</h2>
    			     <br>
        		     <h3>Create a function that will return the x,y position of an element</h3>
        		     <br>
        		     <div class="editor">
         			 <div class="editorText">
         			     <p class="tab1 editorHtml">&lt;script&gt;
             		     <p class="tab2 editorHtml">function checkElementPosition(elemID) {</p>
             		     <p class="tab3 editorHtml"><span class="editorReserveWord">return</span> document.getElementById(elemID).getBoundingClientRect().top +</p>
             		     <p class="tab3 editorHtml">"," +  document.getElementById(elemID).getBoundingClientRect().left;</p>
             		     <p class="tab2 editorHtml">}</p>
             			 <p class="tab1 editorHtml">&lt;/script&gt;
             		 </div>
         			 </div>
    			</section>
    			
                <section>
    			     <h2>Step 10</h2>
    			     <br>
        		     <h3>Use test function from testharness.js</h3>
        		     <h4 style="color:red">test(function, description)</h4>
        		     <br>
        		     <div class="editor">
         			     <div class="editorText">
             			     <br>
             			     <p class="tab1 editorHtml">&lt;script&gt;
                 		     <p class="tab2 editorHtml">...</p>
                 		     <p class="tab2 editorHtml">test(function() {}, "Enter test description here")</p>
                 			 <p class="tab1 editorHtml">&lt;/script&gt;
                 		 </div>
         			 </div>
         			 <br><br><br><br><br><br>
         			 <h4>Each call to test() is a unique test case</h4>
         			 <h4>Pass/Fail is based on return value of function</h4>
         			 <h4>Description is displayed in test results table</h4>
    			</section>    			
    			
                <section>
    			     <h2>Step 11</h2>
    			     <br>
        		     <h3>Use assert_equals to check position of SVG Rect</h3>
        		     <h4 style="color:red">assert_equals(actual, expected, description)</h4>
        		     <h5>NOTE: description is only displayed if assertion failed</h5>
        		     <h5 style="font-style:italic">Additional functions documented in comment section of testharness.js</h5>
        		     <br>
        		     <div class="editor">
         			     <div class="editorText">
                             <br>    			         
         			         <p class="tab2 editorHtml">assert_equals(checkElementPosition(<span class="editorAttrVal">"myRect"</span>), <span class="editorAttrVal">"50,50"</span>,</p>
         			         <p class="tab2 editorHtml"><span class="editorAttrVal">"myRect should be at 50,50"</span>)</p>
             		     </div>
         			 </div>
         			 <br><br><br><br><br>
         			 <div style="text-align:left; line-height:14px; font-size:14px">
         			     <p>actual = assert_equals(checkElementPosition("myRect")</p>
            		     <p>expected = "50,50"</p>
            		     <p>description = "myRect should be at 50,50"</p>
            		</div>
    			</section>
                
                <section>
        		     <h2>Step 12</h2>
        		     <br>
            	     <h3>Call assert_equals function, inline, from test function</h3>
            	     <h4 style="color:red">test(function, description)</h4>
                     <br>
         		     <div class="editor">
          			     <div class="editorText">
                              <br>    			         
          			         <p class="tab2 editorHtml">test(function() {assert_equals(checkElementPosition(<span class="editorAttrVal">"myRect"</span>), <span class="editorAttrVal">"50,50"</span>,</p>
          			         <p class="tab2 editorHtml"><span class="editorAttrVal">"myRect should be at 50,50"</span>)}, <span class="editorAttrVal">"Test transform translate on SVG rect"</span>);</p>
              		     </div>
          			 </div>            	     
                </section>
                
                <section>
                    <h2 style="line-height:10px">Final file</h2>
                    <img src="assets/images/jsTestFinal.png">
                </section>
                
                <section>
        		     <h2>Passing test example</h3>
        		     <br>
            	     <img src="assets/images/passingJSTestExample.png">
                </section>
                
                <section>
    			     <h2>Failing test example</h2>
    			     <br>
             	     <img src="assets/images/failingJSTestExample.png">
                </section>
                
                <section>
    			     <h2>Step 13</h2>
    			     <br>
        		     <h3>Ask someone to review your test</h3>
      		    </section>      
      		    
                <section>
   			     <h2>Step 14</h2>
   			     <br>
       		     <h3>Add your new test file to the repository</h3>
   			     <br>
     			 <div class="console">
 				     <div class="consoleText">
     				     <br><br>
     				     <p>$ cd ~/W3Cdir/test/contributors/yourName</p>
     			         <p>$ hg add</p>
     			         <p>$ hg commit -m "TTWF JavaScript tutorial"</p>
     			         <p>$ hg push</p>
     			     </div>
 			     </div>
      		    </section> 
      		    
      		    <section>
      		        <h2>Questions & Answers</h2>
      		    </section>     		    		     
        </div>

			<!-- The navigational controls UI -->
			<aside class="controls">
				<a class="left" href="#">&#x25C4;</a>
				<a class="right" href="#">&#x25BA;</a>
				<a class="up" href="#">&#x25B2;</a>
				<a class="down" href="#">&#x25BC;</a>
			</aside>

			<!-- Displays presentation progress, max value changes via JS to reflect # of slides -->
			<div class="progress"><span></span></div>
			
		</div>
		
		<script src="js/reveal.js"></script>
		<script src="js/prefixfree.min.js"></script>
		<script src="lib/highlight.js"></script>
		<script>
			// Parse the query string into a key/value object
			var query = {};
			location.search.replace( /[A-Z0-9]+?=(\w*)/gi, function(a) {
				query[ a.split( '=' ).shift() ] = a.split( '=' ).pop();
			} );

			Reveal.initialize({
				// Display controls in the bottom right corner
				controls:true,

				// Display a presentation progress bar
				progress: true,

				// If true; each slide will be pushed to the browser history
				history: true,

				// Flags if mouse wheel navigation should be enabled
				mouseWheel: true,

				// Apply a 3D roll to links on hover
				rollingLinks: true,

				// UI style
				theme: query.theme || 'default', // default/neon

				// Transition style
				transition: query.transition || 'concave' // default/cube/page/concave/linear(2d)
			});

			hljs.initHighlightingOnLoad();
		</script>

	</body>
</html>
